% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZERAH.R
\name{dCertificatBigData}
\alias{dCertificatBigData}
\title{Computes the PDF from samples}
\usage{
dCertificatBigData(X)
}
\arguments{
\item{X}{Numeric vector whose elements are observations of a random variable}
}
\value{
"density" class object
}
\description{
Computes the PDF from samples
}
\examples{
n<-100 # sample number
X<-rnorm(n) # n observation of a normal law
Y<-dCertificatBigData(X) # computing PDF

}
