% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZERAH.R
\name{eCertificatBigData}
\alias{eCertificatBigData}
\title{Computes arithmetic mean from samples}
\usage{
eCertificatBigData(X)
}
\arguments{
\item{X}{Numeric vector whose elements are observations of a random variable}
}
\value{
Numeric vector
}
\description{
Computes arithmetic mean from samples
}
\examples{
n<-100 # sample
X<-rnorm(n) # n observation of a normal law
Y<-eCertificatBigData(X) # computing mean

}
