% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZERAH.R
\name{vCertificatBigData}
\alias{vCertificatBigData}
\title{Computes standart deviation from samples}
\usage{
vCertificatBigData(X)
}
\arguments{
\item{X}{Numeric vector whose elements are observations of a random variable}
}
\value{
Numeric
}
\description{
Computes standart deviation from samples
}
\examples{
n<-100 # sample number
X<-rnorm(n) # n observation of a normal law
Y<-eCertificatBigData(X) # computing sd

}
